Fix handling of -coverprofile
The logic in coverfail currently adds two options if -coverprofile is set:
"-coverprofile=" and the name of the desired profile.  This effectively
disables the output of a coverage profile.  To correct this, the equal
sign is omitted, which will cause the following argument to be interpreted
as a value for the option.