Add support for "-mod"
The go commands take a "-mod" build argument, which is used for
controlling the behavior around dependency handling.  In particular,
"-mod=readonly" can be used to inform the go tools that go.mod should be
considered read-only, which will allow errors to be raised if go.mod
requires edits.  This is obviously a useful property for CI pipelines.
This commit adds support for "-mod" to coverfail.